Kernel: Move HandleTable to a separate file
