+++ src/core/hle/kernel/client_session.h
@@ -6,10 +6,9 @@
 
 #include <memory>
 #include <string>
-
 #include "common/common_types.h"
-
 #include "core/hle/kernel/kernel.h"
+#include "core/hle/result.h"
 
 namespace Kernel {
