+++ src/core/hle/service/apt/apt.h
@@ -4,6 +4,8 @@
 
 #pragma once
 
+#include <vector>
+#include "common/common_funcs.h"
 #include "common/common_types.h"
 #include "common/swap.h"
 #include "core/hle/kernel/kernel.h"
