+++ src/core/hle/service/service.h
@@ -12,7 +12,6 @@
 #include "core/hle/ipc.h"
 #include "core/hle/ipc_helpers.h"
 #include "core/hle/kernel/client_port.h"
-#include "core/hle/kernel/thread.h"
 #include "core/hle/result.h"
 #include "core/memory.h"
